Move examples into examples directory
