+++ examples/example-oc/buf.gen.yaml
