+++ examples/example-otel/buf.gen.yaml
