+++ examples/example-otel/buf.yaml
@@ -1,5 +1,5 @@
 version: v1
-name: buf.build/clly/ocecho
+name: buf.build/clly/otecho
 deps:
   - buf.build/googleapis/googleapis
   - buf.build/clly/telemetry
