+++ examples/example-otel/cmd/server/main.go
